Make improvements
- Get mprotect_test working on aarch64
- Get completion working on python.com repl again
- Improve quality of printvideo.com and printimage.com
- Fix bug in openpty() so examples/script.c works again

+++ libc/str/unicode.h
@@ -32,7 +32,7 @@ struct lconv {
 
 int wcwidth(wchar_t) pureconst;
 int wcswidth(const wchar_t *, size_t) strlenesque;
-int wcsnwidth(const wchar_t *, size_t, int) strlenesque;
+int wcsnwidth(const wchar_t *, size_t, size_t) strlenesque;
 struct lconv *localeconv(void);
 
 COSMOPOLITAN_C_END_
